+++ src/controllers/EventSpecificController.ts
@@ -11,7 +11,7 @@ import { EntitiesEnum, Role } from '../types'
 import { generateRedisKey, generateRedisKeysArray } from '../utils/redisHelper'
 import Context from '../context'
 import { checkUserRole } from '../middlewares'
-import type { EmployeeEntity } from '../entity/EmployeeEntity'
+import type { EmployeeEntity } from '../entity/employees/EmployeeEntity'
 import EmployeeService from '../services/EmployeeService'
 import type { AddressEntity } from '../entity/AddressEntity'
 import { ApiError } from '../middlewares/ApiError'
