+++ src/controllers/MailController.ts
@@ -6,7 +6,7 @@ import { wrapperRequest } from '../utils'
 import { APP_SOURCE } from '..'
 import { ApiError } from '../middlewares/ApiError'
 import AnswerService from '../services/AnswerService'
-import type { EmployeeEntity } from '../entity/EmployeeEntity'
+import type { EmployeeEntity } from '../entity/employees/EmployeeEntity'
 import { firtSendAnswerTemplate } from '../utils/mailJetHelpers'
 import Context from '../context'
 import { isUserAdmin } from '../utils/userHelper'
