+++ src/controllers/NewsletterController.ts
@@ -2,7 +2,7 @@ import type { Request, Response } from 'express'
 import type { EntityManager, Repository } from 'typeorm'
 import { paginator, wrapperRequest } from '../utils'
 import { APP_SOURCE } from '..'
-import { EmployeeEntity } from '../entity/EmployeeEntity'
+import { EmployeeEntity } from '../entity/employees/EmployeeEntity'
 import { UserEntity } from '../entity/UserEntity'
 import { ApiError } from '../middlewares/ApiError'
 import { NewsletterRecipient, newsletterRecipientSearchableFields } from './../entity/NewsletterRecipientEntity'
