+++ src/entity/UserEntity.ts
@@ -3,7 +3,7 @@ import { Role, ThemeEnum } from '../types/'
 import { SubscriptionEnum } from '../types/Subscription'
 import { AddressEntity } from './AddressEntity'
 import { BaseEntity } from './bases/BaseEntity'
-import { EmployeeEntity } from './EmployeeEntity'
+import { EmployeeEntity } from './employees/EmployeeEntity'
 import EventEntity from './EventEntity'
 import { FileEntity } from './FileEntity'
 import { NotificationSubcriptionEntity } from './notifications/NotificationSubscription.entity'
