+++ src/entity/index.ts
@@ -2,7 +2,7 @@ import AnswerEntity from './AnswerEntity'
 import { AddressEntity } from './AddressEntity'
 import { BaseEntity } from './bases/BaseEntity'
 import { BugReportEntity } from './BugReportEntity'
-import { EmployeeEntity } from './EmployeeEntity'
+import { EmployeeEntity } from './employees/EmployeeEntity'
 import EventEntity from './EventEntity'
 import { FileEntity } from './FileEntity'
 import { ImageRightConditionEntity } from './ImageRightConditionEntity'
