+++ src/services/AddressService.ts
@@ -4,7 +4,7 @@ import axios from 'axios'
 import type { AddressCreationServicePayload, GeoCodingResponse } from '../types'
 import EventEntity from '../entity/EventEntity'
 import { AddressEntity } from '../entity/AddressEntity'
-import { EmployeeEntity } from '../entity/EmployeeEntity'
+import { EmployeeEntity } from '../entity/employees/EmployeeEntity'
 import { UserEntity } from '../entity/UserEntity'
 import { isArray } from '../utils/'
 import { useEnv } from '../env'
